+++ src/Armature.Core/src/PatternTree/FindUnitMatches.cs
@@ -19,7 +19,7 @@ namespace Armature.Core
     ///   Moves along the building unit sequence skipping units until it finds the matching unit.
     ///   If it is the unit under construction, returns build actions for it, if no, pass the rest of the sequence to each child and returns merged actions.
     /// </summary>
-    public override WeightedBuildActionBag? GatherBuildActions(ArrayTail<UnitId> unitSequence, long inputWeight)
+    public override WeightedBuildActionBag? GatherBuildActions(ArrayTail<UnitId> unitSequence, int inputWeight)
     {
       var realWeight = inputWeight;
