+++ src/Armature.Core/src/PatternTree/IfLastUnitMatches.cs
@@ -18,7 +18,7 @@ namespace Armature.Core
 
     public override ICollection<IPatternTreeNode> Children => throw new NotSupportedException("This pattern can't contain children");
 
-    public override WeightedBuildActionBag? GatherBuildActions(ArrayTail<UnitId> unitSequence, long inputWeight)
+    public override WeightedBuildActionBag? GatherBuildActions(ArrayTail<UnitId> unitSequence, int inputWeight)
     {
       if(unitSequence.Length > 1) return null;
